Separated out functions for second problem
